refactor: renamed imgproc to matproc
