+++ retinify/src/matproc.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025 Sensui Yagi. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-#include "imgproc.hpp"
+#include "matproc.hpp"
 
 #include "retinify/logging.hpp"
