+++ retinify/src/pipeline.cpp
@@ -1,8 +1,8 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025 Sensui Yagi. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-#include "imgproc.hpp"
 #include "mat.hpp"
+#include "matproc.hpp"
 #include "session.hpp"
 #include "stream.hpp"
